requirements.txt
# =============================================================================
# OCR Receipt System - 패키지 목록
# =============================================================================
# 필요한 라이브러리가 있으면 아래에 추가하세요
# 추가 후: pip install -r requirements.txt
# =============================================================================

streamlit==1.31.0
python-dotenv==1.2.1
supabase==2.27.2
plotly==6.0.1
streamlit-js-eval==0.1.7
google-generativeai==0.8.6
google-cloud-vision==3.12.1
openpyxl==3.1.5
